+++ src/ts/components/audiotracklistbox.ts
@@ -9,6 +9,6 @@ export class AudioTrackListBox extends ListBox {
 
   configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager);
-    new AudioTrackSwitchHandler(player, this);
+    new AudioTrackSwitchHandler(player, this, uimanager);
   }
 }