+++ src/ts/components/subtitleselectbox.ts
@@ -15,6 +15,6 @@ export class SubtitleSelectBox extends SelectBox {
   configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager);
 
-    new SubtitleSwitchHandler(player, this);
+    new SubtitleSwitchHandler(player, this, uimanager);
   }
 }