+++ src/ts/components/volumetogglebutton.ts
@@ -1,5 +1,5 @@
 import {ToggleButton, ToggleButtonConfig} from './togglebutton';
-import {UIInstanceManager} from '../uimanager';
+import { InternalUIConfig, UIInstanceManager } from '../uimanager';
 
 /**
  * A button that toggles audio muting.
@@ -54,6 +54,12 @@ export class VolumeToggleButton extends ToggleButton<ToggleButtonConfig> {
     player.on(player.exports.Event.Muted, muteStateHandler);
     player.on(player.exports.Event.Unmuted, muteStateHandler);
     player.on(player.exports.Event.VolumeChanged, volumeLevelHandler);
+    // Listen to the UI event when components need to update them-self
+    // Will also be triggered on player.exports.Event.SourceLoaded
+    uimanager.getConfig().events.onUpdated.subscribe(() => {
+      muteStateHandler();
+      volumeLevelHandler();
+    });
 
     this.onClick.subscribe(() => {
       if (player.isMuted()) {