+++ src/ts/playerutils.ts
@@ -1,5 +1,6 @@
 import {Event, EventDispatcher, NoArgs} from './eventdispatcher';
 import {BrowserUtils} from './browserutils';
+import { InternalUIConfig, UIInstanceManager, UIManager } from './uimanager';
 
 export namespace PlayerUtils {
 
@@ -93,16 +94,19 @@ export namespace PlayerUtils {
     private player: PlayerAPI;
     private live: boolean;
     private liveChangedEvent = new EventDispatcher<PlayerAPI, LiveStreamDetectorEventArgs>();
+    private uimanager: UIInstanceManager;
 
-    constructor(player: PlayerAPI) {
+    constructor(player: PlayerAPI, uimanager: UIInstanceManager) {
       this.player = player;
+      this.uimanager = uimanager;
       this.live = undefined;
 
       let liveDetector = () => {
         this.detect();
       };
-      // Initialize when source is loaded
-      player.on(player.exports.Event.SourceLoaded, liveDetector);
+      // Listen to the UI event when components need to update them-self
+      // Will also be triggered on player.exports.Event.SourceLoaded
+      this.uimanager.getConfig().events.onUpdated.subscribe(liveDetector);
       // Re-evaluate when playback starts
       player.on(player.exports.Event.Play, liveDetector);